Add a max-age=0 patch for some versions of Safari
Some versions of Safari (at least my version, 7.0.3 OSX) will send
Cache-Control:max-age=0 instead of the usual Cache-Control:no-cache when
refreshing a page, after deleting the cache. It expects the server to
respond with up-to-date data. However, if the server response with a
304, then Safari will display a blank page to the user.

This very simple patch treats "max-age=0" the same as it does
"no-cache": it automatically treats the request as fresh, regardless of
other parameters.

closes #9